// src/types.ts
export interface Song {
    accent: string;
    artist: string;
    cover: string;
    date_created: string;
    date_updated: string;
    duration: string;
    id: number;
    name: string;
    sort: boolean | null;
    status: string;
    top_track: boolean;
    url: string;
    user_created: string;
    user_updated: string
}